Revert "chore(dataset-versioning): add dual-write for dataset-item-events (#1…"
This reverts commit e7fc1c940ed5b651aaf689d5cc91684a6f82b894.
