+++ packages/shared/src/server/services/DatasetService/index.ts
@@ -1,3 +1,4 @@
+export * from "./DatasetItemManager";
 export * from "./types";
 export * from "./DatasetItemValidator";
 export * from "./DatasetSchemaValidator";
