+++ web/src/pages/api/public/dataset-items/index.ts
@@ -8,8 +8,12 @@ import {
   PostDatasetItemsV1Response,
   transformDbDatasetItemToAPIDatasetItem,
 } from "@/src/features/public-api/types/datasets";
-import { LangfuseNotFoundError, Prisma } from "@langfuse/shared";
-import { logger, upsertDatasetItem } from "@langfuse/shared/src/server";
+import {
+  LangfuseNotFoundError,
+  InvalidRequestError,
+  Prisma,
+} from "@langfuse/shared";
+import { logger, DatasetItemManager } from "@langfuse/shared/src/server";
 import { auditLog } from "@/src/features/audit-logs/auditLog";
 
 export const config = {
@@ -39,7 +43,7 @@ export default withMiddlewares({
       } = body;
 
       try {
-        const datasetItem = await upsertDatasetItem({
+        const result = await DatasetItemManager.upsertItem({
           projectId: auth.scope.projectId,
           datasetName: datasetName,
           datasetItemId: id ?? undefined,
@@ -52,19 +56,24 @@ export default withMiddlewares({
           normalizeOpts: { sanitizeControlChars: true },
           validateOpts: { normalizeUndefinedToNull: !!id ? false : true },
         });
+        if (!result.success) {
+          throw new InvalidRequestError(
+            `Dataset item validation failed: ${result.message}`,
+          );
+        }
 
         await auditLog({
           action: "create",
           resourceType: "datasetItem",
-          resourceId: datasetItem.id,
+          resourceId: result.datasetItem.id,
           projectId: auth.scope.projectId,
           orgId: auth.scope.orgId,
           apiKeyId: auth.scope.apiKeyId,
-          after: datasetItem,
+          after: result.datasetItem,
         });
 
         return transformDbDatasetItemToAPIDatasetItem({
-          ...datasetItem,
+          ...result.datasetItem,
           datasetName: datasetName,
         });
       } catch (e) {
